pyproject.toml
[project]
name = "secretsantabot"
version = "1.20251028.1"
description = "A Secret Santa gift exchange bot with Flask"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.1.0",
    "flask-limiter>=3.5.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["data*", "dev-tools*", "docs*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \.venv
  | venv
  | data
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["data", ".venv", "venv"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]
